Rename `rtc` to `Clock`
